Move datasource files to own package
